Migrate some tests to cucumber (#2682)
- Refactored cucumber tests to better separate solver & modeler steps
- Added table support to check annual solver results
- Added the ability to create temporary copies of studies before
runnning the test, in order to be able to modify the study before
running the simulation
- Migrated some short-tests
- Added the ability to skip, in run-study-tests, scenarios that are
tested in cucumber

---------


(cherry picked from commit 734e0e637377150a24a9046f2c2d40c7762e6704)
