+++ src/tests/cucumber/features/solver-features/hybrid_studies.feature
@@ -2,13 +2,13 @@ Feature: hybrid (simulator+modeler) studies
 
   @fast @short
   Scenario: 001 One node - passive + modeler test 1_1
-    Given the study path is "Antares_Simulator_Tests_NR/hybrid/001 One node - passive"
+    Given the solver study path is "Antares_Simulator_Tests_NR/hybrid/001 One node - passive"
     When I run antares simulator
     Then the simulation succeeds
 
   @fast @short
   Scenario: 002 Thermal fleet - Base + modeler test 1_1
-    Given the study path is "Antares_Simulator_Tests_NR/hybrid/002 Thermal fleet - Base"
+    Given the solver study path is "Antares_Simulator_Tests_NR/hybrid/002 Thermal fleet - Base"
     When I run antares simulator
     Then the simulation succeeds
     And the annual system cost is 2.729e+7
