+++ src/tests/cucumber/features/steps/common_steps/assertions.py
@@ -1,5 +1,5 @@
 # Custom assertions
 
-def assert_double_close(expected, actual, relative_tolerance):
+def assert_double_close(expected, actual, relative_tolerance, message_prefix="Compared"):
     assert abs((actual - expected) / max(1e-6, expected)) <= relative_tolerance, \
-        f"Values are not close: expected = {expected} ; actual = {actual}"
+     f"{message_prefix} values are not close: expected = {expected} ; actual = {actual}"
