+++ src/tests/run-study-tests/requirements.txt
@@ -1,0 +1,4 @@
+pytest
+jsonschema
+pandas
+numpy
